Trait chalk_engine::context::ContextOps

pub trait ContextOps<C: Context>: Sized + Clone + Debug + AggregateOps<C> {
    fn is_coinductive(&self, goal: &C::UCanonicalGoalInEnvironment) -> bool;

    fn instantiate_ucanonical_goal<R>(
        &self, 
        arg: &C::UCanonicalGoalInEnvironment, 
        op: impl WithInstantiatedUCanonicalGoal<C, Output = R>
    ) -> R;

    fn instantiate_ex_clause<R>(
        &self, 
        num_universes: usize, 
        canonical_ex_clause: &C::CanonicalExClause, 
        op: impl WithInstantiatedExClause<C, Output = R>
    ) -> R;

    fn inference_normalized_subst_from_ex_clause(
        canon_ex_clause: &C::CanonicalExClause
    ) -> &C::InferenceNormalizedSubst;

    fn inference_normalized_subst_from_subst(
        canon_ex_clause: &C::CanonicalConstrainedSubst
    ) -> &C::InferenceNormalizedSubst;

    fn empty_constraints(ccs: &C::CanonicalConstrainedSubst) -> bool;

    fn canonical(
        u_canon: &C::UCanonicalGoalInEnvironment
    ) -> &C::CanonicalGoalInEnvironment;

    fn is_trivial_substitution(
        u_canon: &C::UCanonicalGoalInEnvironment, 
        canonical_subst: &C::CanonicalConstrainedSubst
    ) -> bool;

    fn num_universes(_: &C::UCanonicalGoalInEnvironment) -> usize;

    fn map_goal_from_canonical(
        _: &C::UniverseMap, 
        value: &C::CanonicalGoalInEnvironment
    ) -> C::CanonicalGoalInEnvironment;

    fn map_subst_from_canonical(
        _: &C::UniverseMap, 
        value: &C::CanonicalConstrainedSubst
    ) -> C::CanonicalConstrainedSubst;
}

Required Methods

fn is_coinductive(&self, goal: &C::UCanonicalGoalInEnvironment) -> bool

True if this is a coinductive goal -- e.g., proving an auto trait.

fn instantiate_ucanonical_goal<R>(
&self,
arg: &C::UCanonicalGoalInEnvironment,
op: impl WithInstantiatedUCanonicalGoal<C, Output = R>
) -> R

Create an inference table for processing a new goal and instantiate that goal in that context, returning "all the pieces".

More specifically: given a u-canonical goal arg, creates a new inference table T and populates it with the universes found in arg. Then, creates a substitution S that maps each bound variable in arg to a fresh inference variable from T. Returns:

the table T

the substitution S

the environment and goal found by substitution S into arg

fn instantiate_ex_clause<R>(
&self,
num_universes: usize,
canonical_ex_clause: &C::CanonicalExClause,
op: impl WithInstantiatedExClause<C, Output = R>
) -> R

fn inference_normalized_subst_from_ex_clause(
canon_ex_clause: &C::CanonicalExClause
) -> &C::InferenceNormalizedSubst

Extracts the inner normalized substitution from a canonical ex-clause.

fn inference_normalized_subst_from_subst(
canon_ex_clause: &C::CanonicalConstrainedSubst
) -> &C::InferenceNormalizedSubst

Extracts the inner normalized substitution from a canonical constraint subst.

fn empty_constraints(ccs: &C::CanonicalConstrainedSubst) -> bool

True if this solution has no region constraints.

fn canonical(
u_canon: &C::UCanonicalGoalInEnvironment
) -> &C::CanonicalGoalInEnvironment

fn is_trivial_substitution(
u_canon: &C::UCanonicalGoalInEnvironment,
canonical_subst: &C::CanonicalConstrainedSubst
) -> bool

fn num_universes(_: &C::UCanonicalGoalInEnvironment) -> usize

fn map_goal_from_canonical(
_: &C::UniverseMap,
value: &C::CanonicalGoalInEnvironment
) -> C::CanonicalGoalInEnvironment

Convert a goal G from the canonical universes into our local universes. This will yield a goal G' that is the same but for the universes of universally quantified names.

fn map_subst_from_canonical(
_: &C::UniverseMap,
value: &C::CanonicalConstrainedSubst
) -> C::CanonicalConstrainedSubst

Convert a substitution from the canonical universes into our local universes. This will yield a substitution S' that is the same but for the universes of universally quantified names.