Trait chalk_engine::context::InferenceTable

pub trait InferenceTable<C: Context, I: Context>: ResolventOps<C, I> + TruncateOps<C, I> + UnificationOps<C, I> {
    fn into_hh_goal(&mut self, goal: I::Goal) -> HhGoal<I>;

    fn add_clauses(
        &mut self, 
        env: &I::Environment, 
        clauses: I::ProgramClauses
    ) -> I::Environment;

    fn into_goal(&self, domain_goal: I::DomainGoal) -> I::Goal;

    fn cannot_prove(&self) -> I::Goal;
}

An "inference table" contains the state to support unification and other operations on terms.

Required Methods

fn into_hh_goal(&mut self, goal: I::Goal) -> HhGoal<I>

Convert the context's goal type into the HhGoal type that the SLG solver understands. The expectation is that the context's goal type has the same set of variants, but with different names and a different setup. If you inspect HhGoal, you will see that this is a "shallow" or "lazy" conversion -- that is, we convert the outermost goal into an HhGoal, but the goals contained within are left as context goals.

fn add_clauses(
&mut self,
env: &I::Environment,
clauses: I::ProgramClauses
) -> I::Environment

fn into_goal(&self, domain_goal: I::DomainGoal) -> I::Goal

Upcast this domain goal into a more general goal.

fn cannot_prove(&self) -> I::Goal

Create a "cannot prove" goal (see HhGoal::CannotProve).