Module chalk_engine::logic

Structs

SolveState

Enums

NoRemainingSubgoalsResult
This is returned on_no_remaining_subgoals

RootSearchFail
The different ways that a root search (which potentially pursues many strands) can fail. A root search is one that begins with an empty stack.

SubGoalSelection
This is returned when we try to select a subgoal for a strand.

Type Definitions

RootSearchResult