Struct chalk_engine::tables::Tables

pub(crate) struct Tables<I: Interner> {
    table_indices: FxHashMap<UCanonical<InEnvironment<Goal<I>>>, TableIndex>,
    tables: Vec<Table<I>>,
}

See Forest.

Fields
table_indices: FxHashMap<UCanonical<InEnvironment<Goal<I>>>, TableIndex>

Maps from a canonical goal to the index of its table.

tables: Vec<Table<I>>

Table: as described above, stores the key information for each tree in the forest.

Implementations

impl<I: Interner> Tables<I>

pub(crate) fn new() -> Tables<I>

pub(crate) fn next_index(&self) -> TableIndex

The index that will be given to the next table to be inserted.

pub(crate) fn insert(&mut self, table: Table<I>) -> TableIndex

pub(crate) fn index_of(
&self,
literal: &UCanonical<InEnvironment<Goal<I>>>
) -> Option<TableIndex>