Trait chalk_engine::context::ContextOps

pub trait ContextOps<C: Context>: Sized + Clone + Debug + AggregateOps<C> {
    fn is_coinductive(&self, goal: &C::UCanonicalGoalInEnvironment) -> bool;

    fn identity_constrained_subst(
        &self, 
        goal: &C::UCanonicalGoalInEnvironment
    ) -> C::CanonicalConstrainedSubst;

    fn instantiate_ucanonical_goal<R>(
        &self, 
        arg: &C::UCanonicalGoalInEnvironment, 
        op: impl WithInstantiatedUCanonicalGoal<C, Output = R>
    ) -> R;

    fn instantiate_ex_clause<R>(
        &self, 
        num_universes: usize, 
        canonical_ex_clause: &C::CanonicalExClause, 
        op: impl WithInstantiatedExClause<C, Output = R>
    ) -> R;
}

Required methods

fn is_coinductive(&self, goal: &C::UCanonicalGoalInEnvironment) -> bool

True if this is a coinductive goal -- e.g., proving an auto trait.

fn identity_constrained_subst(
&self,
goal: &C::UCanonicalGoalInEnvironment
) -> C::CanonicalConstrainedSubst

Returns a identity substitution.

fn instantiate_ucanonical_goal<R>(
&self,
arg: &C::UCanonicalGoalInEnvironment,
op: impl WithInstantiatedUCanonicalGoal<C, Output = R>
) -> R

Create an inference table for processing a new goal and instantiate that goal in that context, returning "all the pieces".

More specifically: given a u-canonical goal arg, creates a new inference table T and populates it with the universes found in arg. Then, creates a substitution S that maps each bound variable in arg to a fresh inference variable from T. Returns:

the table T

the substitution S

the environment and goal found by substitution S into arg

fn instantiate_ex_clause<R>(
&self,
num_universes: usize,
canonical_ex_clause: &C::CanonicalExClause,
op: impl WithInstantiatedExClause<C, Output = R>
) -> R

Loading content...