Trait chalk_engine::context::UnificationOps

pub trait UnificationOps<C: Context> {
    fn instantiate_binders_universally(
        &mut self, 
        interner: &C::Interner, 
        arg: &C::BindersGoal
    ) -> C::Goal;

    fn instantiate_binders_existentially(
        &mut self, 
        interner: &C::Interner, 
        arg: &C::BindersGoal
    ) -> C::Goal;

    fn debug_ex_clause<'v>(
        &mut self, 
        interner: &C::Interner, 
        value: &'v ExClause<C>
    ) -> Box<dyn Debug + 'v>;

    fn fully_canonicalize_goal(
        &mut self, 
        interner: &C::Interner, 
        value: &C::GoalInEnvironment
    ) -> (C::UCanonicalGoalInEnvironment, C::UniverseMap);

    fn canonicalize_ex_clause(
        &mut self, 
        interner: &C::Interner, 
        value: &ExClause<C>
    ) -> C::CanonicalExClause;

    fn canonicalize_constrained_subst(
        &mut self, 
        interner: &C::Interner, 
        subst: C::Substitution, 
        constraints: Vec<C::RegionConstraint>
    ) -> C::CanonicalConstrainedSubst;

    fn canonicalize_answer_subst(
        &mut self, 
        interner: &C::Interner, 
        subst: C::Substitution, 
        constraints: Vec<C::RegionConstraint>, 
        delayed_subgoals: Vec<C::GoalInEnvironment>
    ) -> C::CanonicalAnswerSubst;

    fn invert_goal(
        &mut self, 
        interner: &C::Interner, 
        value: &C::GoalInEnvironment
    ) -> Option<C::GoalInEnvironment>;

    fn unify_parameters_into_ex_clause(
        &mut self, 
        interner: &C::Interner, 
        environment: &C::Environment, 
        variance: C::Variance, 
        a: &C::Parameter, 
        b: &C::Parameter, 
        ex_clause: &mut ExClause<C>
    ) -> Fallible<()>;
}

Methods for unifying and manipulating terms and binders.

Required methods

fn instantiate_binders_universally(
&mut self,
interner: &C::Interner,
arg: &C::BindersGoal
) -> C::Goal

fn instantiate_binders_existentially(
&mut self,
interner: &C::Interner,
arg: &C::BindersGoal
) -> C::Goal

fn debug_ex_clause<'v>(
&mut self,
interner: &C::Interner,
value: &'v ExClause<C>
) -> Box<dyn Debug + 'v>

fn fully_canonicalize_goal(
&mut self,
interner: &C::Interner,
value: &C::GoalInEnvironment
) -> (C::UCanonicalGoalInEnvironment, C::UniverseMap)

fn canonicalize_ex_clause(
&mut self,
interner: &C::Interner,
value: &ExClause<C>
) -> C::CanonicalExClause

fn canonicalize_constrained_subst(
&mut self,
interner: &C::Interner,
subst: C::Substitution,
constraints: Vec<C::RegionConstraint>
) -> C::CanonicalConstrainedSubst

fn canonicalize_answer_subst(
&mut self,
interner: &C::Interner,
subst: C::Substitution,
constraints: Vec<C::RegionConstraint>,
delayed_subgoals: Vec<C::GoalInEnvironment>
) -> C::CanonicalAnswerSubst

fn invert_goal(
&mut self,
interner: &C::Interner,
value: &C::GoalInEnvironment
) -> Option<C::GoalInEnvironment>

fn unify_parameters_into_ex_clause(
&mut self,
interner: &C::Interner,
environment: &C::Environment,
variance: C::Variance,
a: &C::Parameter,
b: &C::Parameter,
ex_clause: &mut ExClause<C>
) -> Fallible<()>

First unify the parameters, then add the residual subgoals as new subgoals of the ex-clause. Also add region constraints.

If the parameters fail to unify, then Error is returned

Loading content...