Struct chalk_engine::forest::Forest

pub struct Forest<C: Context> {
    context: C,
    pub(crate) tables: Tables<C>,
    pub(crate) clock: TimeStamp,
}

Fields
context: Ctables: Tables<C>clock: TimeStamp

This is a clock which always increases. It is incremented every time a new subgoal is followed. This effectively gives us way to track what depth and loop a table or strand was last followed.

Methods

impl<C: Context> Forest<C>

pub fn new(context: C) -> Self

pub fn context(&self) -> &C

Gives access to self.context. In fact, the SLG solver doesn't ever use self.context for anything, and only cares about the associated types and methods defined on it. But the creator of the forest can use the context field to store configuration info (e.g., in chalk, we store the max size of a term in here).

pub(crate) fn increment_clock(&mut self) -> TimeStamp

fn iter_answers<'f>(
&'f mut self,
context: &'f impl ContextOps<C>,
goal: &C::UCanonicalGoalInEnvironment
) -> impl AnswerStream<C> + 'f

Returns a "solver" for a given goal in the form of an iterator. Each time you invoke next, it will do the work to extract one more answer. These answers are cached in between invocations. Invoking next fewer times is preferable =)

pub fn solve(
&mut self,
context: &impl ContextOps<C>,
goal: &C::UCanonicalGoalInEnvironment,
should_continue: impl Fn() -> bool
) -> Option<C::Solution>

Solves a given goal, producing the solution. This will do only as much work towards goal as it has to (and that works is cached for future attempts).

pub fn solve_multiple(
&mut self,
context: &impl ContextOps<C>,
goal: &C::UCanonicalGoalInEnvironment,
f: impl FnMut(SubstitutionResult<C::CanonicalConstrainedSubst>, bool) -> bool
) -> bool

Solves a given goal, producing the solution. This will do only as much work towards goal as it has to (and that works is cached for future attempts). Calls provided function f to iterate over multiple solutions until the function return false.

impl<C: Context> Forest<C>

pub(super) fn root_answer(
&mut self,
context: &impl ContextOps<C>,
table: TableIndex,
answer_index: AnswerIndex
) -> Result<CompleteAnswer<C>, RootSearchFail>

Returns an answer with a given index for the given table. This may require activating a strand and following it. It returns Ok(answer) if they answer is available and otherwise a RootSearchFail result.

pub(super) fn any_future_answer(
&self,
table: TableIndex,
answer: AnswerIndex,
test: impl FnMut(&C::InferenceNormalizedSubst) -> bool
) -> bool

pub(crate) fn answer(
&self,
table: TableIndex,
answer: AnswerIndex
) -> &Answer<C>

fn canonicalize_strand(
context: &impl ContextOps<C>,
strand: Strand<C>
) -> CanonicalStrand<C>

fn canonicalize_strand_from(
context: &impl ContextOps<C>,
infer: &mut dyn InferenceTable<C>,
ex_clause: &ExClause<C>,
selected_subgoal: Option<SelectedSubgoal<C>>,
last_pursued_time: TimeStamp
) -> CanonicalStrand<C>

fn get_or_create_table_for_subgoal(
&mut self,
context: &impl ContextOps<C>,
infer: &mut dyn InferenceTable<C>,
subgoal: &Literal<C>
) -> Option<(TableIndex, C::UniverseMap)>

Given a subgoal, converts the literal into u-canonical form and searches for an existing table. If one is found, it is returned, but otherwise a new table is created (and populated with its initial set of strands).

Returns None if the literal cannot be converted into a table -- for example, this can occur when we have selected a negative literal with free existential variables, in which case the execution is said to "flounder".

In terms of the NFTD paper, creating a new table corresponds to the New Subgoal step as well as the Program Clause Resolution steps.

pub(crate) fn get_or_create_table_for_ucanonical_goal(
&mut self,
context: &impl ContextOps<C>,
goal: C::UCanonicalGoalInEnvironment
) -> TableIndex

Given a u-canonical goal, searches for an existing table. If one is found, it is returned, but otherwise a new table is created (and populated with its initial set of strands).

In terms of the NFTD paper, creating a new table corresponds to the New Subgoal step as well as the Program Clause Resolution steps.

fn push_initial_strands(
&mut self,
context: &impl ContextOps<C>,
table: TableIndex
)

When a table is first created, this function is invoked to create the initial set of strands. If the table represents a domain goal, these strands are created from the program clauses as well as the clauses found in the environment. If the table represents a non-domain goal, such as for<T> G etc, then simplify_hh_goal is invoked to create a strand that breaks the goal down.

In terms of the NFTD paper, this corresponds to the Program Clause Resolution step being applied eagerly, as many times as possible.

fn push_initial_strands_instantiated(
&mut self,
context: &impl ContextOps<C>,
table: TableIndex,
infer: C::InferenceTable,
subst: C::Substitution,
environment: C::Environment,
goal: C::Goal
)

fn abstract_positive_literal(
context: &impl ContextOps<C>,
infer: &mut dyn InferenceTable<C>,
subgoal: &C::GoalInEnvironment
) -> Option<(C::UCanonicalGoalInEnvironment, C::UniverseMap)>

Given a selected positive subgoal, applies the subgoal abstraction function to yield the canonical form that will be used to pick a table. Typically, this abstraction has no effect, and hence we are simply returning the canonical form of subgoal; but if the subgoal is getting too big, we return None, which causes the subgoal to flounder.

fn abstract_negative_literal(
context: &impl ContextOps<C>,
infer: &mut dyn InferenceTable<C>,
subgoal: &C::GoalInEnvironment
) -> Option<(C::UCanonicalGoalInEnvironment, C::UniverseMap)>

Given a selected negative subgoal, the subgoal is "inverted" (see InferenceTable<C>::invert) and then potentially truncated (see abstract_positive_literal). The result subgoal is canonicalized. In some cases, this may return None and hence fail to yield a useful result, for example if free existential variables appear in subgoal (in which case the execution is said to "flounder").

impl<C: Context> Forest<C>

pub(super) fn simplify_hh_goal(
context: &impl ContextOps<C>,
infer: &mut dyn InferenceTable<C>,
subst: C::Substitution,
initial_environment: C::Environment,
initial_hh_goal: HhGoal<C>
) -> Fallible<ExClause<C>>

Simplifies an HH goal into a series of positive domain goals and negative HH goals. This operation may fail if the HH goal includes unifications that cannot be completed.