Struct chalk_engine::slg::MayInvalidate

struct MayInvalidate<I> {
    interner: I,
}

Fields
interner: I
Implementations

impl<I: Interner> MayInvalidate<I>

fn aggregate_generic_args(
&mut self,
new: &GenericArg<I>,
current: &GenericArg<I>
) -> bool

fn aggregate_tys(&mut self, new: &Ty<I>, current: &Ty<I>) -> bool

Returns true if the two types could be unequal.

fn aggregate_lifetimes(&mut self, _: &Lifetime<I>, _: &Lifetime<I>) -> bool

Returns true if the two consts could be unequal.

fn aggregate_consts(&mut self, new: &Const<I>, current: &Const<I>) -> bool

Returns true if the two consts could be unequal.

fn aggregate_placeholders(
&mut self,
new: &PlaceholderIndex,
current: &PlaceholderIndex
) -> bool

fn aggregate_projection_tys(
&mut self,
new: &ProjectionTy<I>,
current: &ProjectionTy<I>
) -> bool

fn aggregate_opaque_ty_tys(
&mut self,
new: &OpaqueTy<I>,
current: &OpaqueTy<I>
) -> bool

fn aggregate_name_and_substs<N>(
&mut self,
new_name: N,
new_substitution: &Substitution<I>,
current_name: N,
current_substitution: &Substitution<I>
) -> bool where
N: Copy + Eq + Debug,